fix: manifest 경로
