+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const i=(i,n)=>(i=new URL(i+".js",n).href,s[i]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=s,document.head.appendChild(e)}else e=i,importScripts(i),s()})).then((()=>{let e=s[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e})));self.define=(n,a)=>{const t=e||("document"in self?document.currentScript.src:"")||location.href;if(s[t])return;let c={};const r=e=>i(e,t),o={module:{uri:t},exports:c,require:r};s[t]=Promise.all(n.map((e=>o[e]||r(e)))).then((e=>(a(...e),c)))}}define(["./workbox-4754cb34"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"319855ff32e076f88286b07eab457088"},{url:"/_next/static/EFmiv9pMCeJzdROrX7kzN/_buildManifest.js",revision:"c155cce658e53418dec34664328b51ac"},{url:"/_next/static/EFmiv9pMCeJzdROrX7kzN/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/117-b92039adc5a4498c.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/8e1d74a4-d524d3cde1f03954.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/972-2ad8ee1802622980.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/app/_not-found/page-4eb93a37eb147fac.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/app/d-day/page-2d4f9e21aa53f8e3.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/app/layout-ef239bd1a1476f9b.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/app/offline/page-7312570d77cafb7f.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/app/page-18a65119f69904b4.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/app/percentage-cal/page-91f436b78d958f37.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/app/working-day/page-1480b733b070bcaf.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/fd9d1056-2fce5ef1b8d9778d.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/framework-f66176bb897dc684.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/main-app-26f5a19423862029.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/main-e5dad1f268ffe2d4.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/pages/_app-72b849fbd24ac258.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/pages/_error-7ba65e1336b92748.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-8658fdbc42bfbbae.js",revision:"EFmiv9pMCeJzdROrX7kzN"},{url:"/_next/static/css/fd770f0d86163c29.css",revision:"fd770f0d86163c29"},{url:"/ads.txt",revision:"20e2f74bd1ef4ec5306aeb89ab232d5a"},{url:"/images/favicon.png",revision:"38f2e20c83081ff371bd7a648b367015"},{url:"/images/icons/icon-128x128.png",revision:"0523bc9937d095647bb75f761178b9f6"},{url:"/images/icons/icon-192x192.png",revision:"8446ccbe52d01ed21fd0266b69f87f62"},{url:"/images/icons/icon-512x512.png",revision:"7254db1ad2a72b2a57e5e71c9d5026f0"},{url:"/images/icons/icon.jpeg",revision:"42889277b3dcf10236befcb20b470e5a"},{url:"/images/preview.jpeg",revision:"42889277b3dcf10236befcb20b470e5a"},{url:"/manifest.json",revision:"a2fd5a4e6d18473f3b1d04eccfa58f4b"},{url:"/offline.html",revision:"c4b30cb3e43da632affdcfe905b73cfb"},{url:"/robots.txt",revision:"a1eae90576c7acfaf2c14fb3a7ca2ebb"},{url:"/sitemap.xml",revision:"eab29f788c6f357823f2530cf1781303"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:i,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
+if(!self.define){let e,s={};const n=(n,i)=>(n=new URL(n+".js",i).href,s[n]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=n,e.onload=s,document.head.appendChild(e)}else e=n,importScripts(n),s()})).then((()=>{let e=s[n];if(!e)throw new Error(`Module ${n} didn’t register its module`);return e})));self.define=(i,c)=>{const a=e||("document"in self?document.currentScript.src:"")||location.href;if(s[a])return;let t={};const r=e=>n(e,a),o={module:{uri:a},exports:t,require:r};s[a]=Promise.all(i.map((e=>o[e]||r(e)))).then((e=>(c(...e),t)))}}define(["./workbox-4754cb34"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"723e10bcdc051ded24b261f1a5ef163b"},{url:"/_next/static/chunks/117-e881abe2c0a0c050.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/972-25f2910cc6db301c.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/app/_not-found/page-d1916c0b16ebe328.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/app/d-day/page-0ef9d0b73efd8325.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/app/layout-5a5c33cf048c9c26.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/app/offline/page-7312570d77cafb7f.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/app/page-18a65119f69904b4.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/app/percentage-cal/page-4dcb198cd40b374b.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/app/working-day/page-f5a78123be07d408.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/fd9d1056-7d9b70e9456442b8.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/framework-f66176bb897dc684.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/main-90924605306af99f.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/main-app-26f5a19423862029.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/pages/_app-72b849fbd24ac258.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/pages/_error-7ba65e1336b92748.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-8658fdbc42bfbbae.js",revision:"iqCMsnz71vmczi4MPGm6X"},{url:"/_next/static/css/fd770f0d86163c29.css",revision:"fd770f0d86163c29"},{url:"/_next/static/iqCMsnz71vmczi4MPGm6X/_buildManifest.js",revision:"c155cce658e53418dec34664328b51ac"},{url:"/_next/static/iqCMsnz71vmczi4MPGm6X/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/ads.txt",revision:"20e2f74bd1ef4ec5306aeb89ab232d5a"},{url:"/images/favicon.png",revision:"38f2e20c83081ff371bd7a648b367015"},{url:"/images/icons/icon-128x128.png",revision:"0523bc9937d095647bb75f761178b9f6"},{url:"/images/icons/icon-192x192.png",revision:"8446ccbe52d01ed21fd0266b69f87f62"},{url:"/images/icons/icon-512x512.png",revision:"7254db1ad2a72b2a57e5e71c9d5026f0"},{url:"/images/icons/icon.jpeg",revision:"42889277b3dcf10236befcb20b470e5a"},{url:"/images/preview.jpeg",revision:"42889277b3dcf10236befcb20b470e5a"},{url:"/offline.html",revision:"c4b30cb3e43da632affdcfe905b73cfb"},{url:"/robots.txt",revision:"a1eae90576c7acfaf2c14fb3a7ca2ebb"},{url:"/sitemap.xml",revision:"eab29f788c6f357823f2530cf1781303"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:n,state:i})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
