+++ src/app/layout.tsx
@@ -1,9 +1,28 @@
-"use client";
-
 import Footer from "@/components/Footer";
-import Script from "next/script";
+import { Metadata } from "next";
 import "./globals.css";
-import { useEffect } from "react";
+
+export const metadata: Metadata = {
+  metadataBase: new URL("https://every-calculator.vercel.app"),
+  title: "다양한 계산기 모음",
+  description: "다양한 계산기를 간단하고 편리하게 활용해 보세요!",
+  manifest: "/manifest.webmanifest",
+  icons: {
+    icon: "/images/favicon.png",
+  },
+  openGraph: {
+    type: "website",
+    url: "/",
+    title: "다양한 계산기 모음",
+    description: "다양한 계산기를 간단하고 편리하게 활용해 보세요!",
+    images: [
+      {
+        url: "/images/preview.jpeg",
+        alt: "다양한 계산기 모음 미리보기 이미지",
+      },
+    ],
+  },
+};
 
 export default function RootLayout({
   children,
