+++ src/app/page.tsx
@@ -1,26 +1,4 @@
 import Link from "next/link";
-import { Metadata } from "next";
-
-export const metadata: Metadata = {
-  metadataBase: new URL("https://every-calculator.vercel.app"),
-  title: "다양한 계산기 모음",
-  description: "다양한 계산기를 간단하고 편리하게 활용해 보세요!",
-  icons: {
-    icon: "/images/favicon.png",
-  },
-  openGraph: {
-    type: "website",
-    url: "/",
-    title: "다양한 계산기 모음",
-    description: "다양한 계산기를 간단하고 편리하게 활용해 보세요!",
-    images: [
-      {
-        url: "/images/preview.jpeg",
-        alt: "다양한 계산기 모음 미리보기 이미지",
-      },
-    ],
-  },
-};
 
 const Home = () => {
   return (
